ISSUE-#26: refactor/change_directory_name スペルミス修正
infrastructuresに修正
