+++ trend-collector/trendcollector/libs/__init__.py
@@ -1,4 +1,4 @@
 #!/usr/bin/env /Users/yasui/Practice/trend_collector/trend-collector/venv/bin/python
 
-from .infrastractures import *
+from .infrastructures import *
 from .services.client import *
