+++ trend-collector/trendcollector/libs/infrastructures/__init__.py
