+++ trend-collector/trendcollector/libs/infrastructures/client/__init__.py
