+++ trend-collector/trendcollector/libs/infrastructures/conf.py
