+++ trend-collector/trendcollector/libs/infrastructures/logger.py
