+++ trend-collector/trendcollector/libs/infrastructures/repositories/__init__.py
