+++ trend-collector/trendcollector/libs/infrastructures/repositories/schemas.py
