+++ trend-collector/trendcollector/libs/infrastructures/repositories/trend_repository.py
