+++ trend-collector/trendcollector/libs/infrastructures/repositories/twitter_account.py
